LayoutTests/imported/w3c/web-platform-tests/wai-aria/role/tree-roles-expected.txt

Tests tree and related roles.

x
x
x
x
x	x	x
x	x	x
x	x	x
x	x	x

PASS role is tree
PASS role is treeitem (in tree)
PASS role is group (in treeitem)
PASS role is treeitem (in group, in treeitem)
PASS role is treegrid
PASS role is row (in treegrid)
PASS role is gridcell (in row, in treegrid)
